wip: implemented USB feature reports alongside the CBOR logic
This commit is primarily intended to preserve the WIP state of this
branch, where CBOR based transmission over a CDC UART has been
implemented, along with some basic infrastructure for USB feature
reports.

It is likely future development will focus on HID feature reports,
so this branch primarily intends to preserve the CBOR code in case it is
useful in the future.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>