[noup] feat: added support for watchdog to ZMK firmware
Added support for a basic watchdog to ZMK firmware. Watchdog
functionality is currently implemented using a kernel timer, which
services the watchdog. This will allow the firmware to automatically
reboot if a fault is encountered that stops execution

However, this feature does not currently monitor keyboard functionality
itself- that is, the key scan driver could crash without affecting the
watchdog timer, and the system would not reset. A potential improvement
here would be to integrate wdt_feed() calls throughout the program flow.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>