[noup] fix animation ripple buffer wrap handling
When keypresses come in quickly, ripple animation buffer will overflow.
Since this buffer is a FIFO, the detection logic needs a method to
distinguish between an overflow and empty. Add a flag to detect this,
and force the render logic to run so that the FIFO will drain.

This fixes an issue where rapid keypresses could cause the ripple
animation to stop rendering until a new animation was cycled in.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>